feat: add AddMaterialModal and MaterialsTable components with form for uploading course materials
